Add Mauritius.
It is a special case because the fields have some special definition,
like first four chars are [A-Z] and the next two are [0-9].

In local2iban, I need to only use the right length of the input string
for the % calls, I also need to dup it so that I don't change the
arguments I was called with.

In iban2local, I need to move the removing of the leading 0s at the end
so that I don't remove some from the middle of a field.